Write to a +Drive, after guessing cost four power cycles and a capture cost minutes
Single-chunk +Drive writes now work on an Analog Four, verified by writing three
files to /soundbanks/P and reading each back.

The layouts were captured off Elektron's own Transfer 1.10.4 with a CoreMIDI
spy, not derived. Two days of putting candidate 0x58 bodies to the box produced
three clean refusals, three hangs and no working write — and on this box a body
it cannot parse takes down the whole SysEx API, so each hang cost a power cycle
while a DT2 and DN2 on the same bus answered normally. That hazard is in
probe_drive_write.rs's header, because "the box went offline" is the wrong
conclusion to draw from it.

  0x57 WriteOpen   u32be total-len, path\0            -> ok, u32be fd
  0x58 Write       u32be fd, u32be offset, u32be crc,
                   u32be chunk-len, data              -> ok, fd, offset, taken
  0x59 WriteClose  u32be fd, u32be 1                  -> ok, fd, total-len

Three details no sweep would have reached: the 0x58 body has four u32 fields in
an order no gen-1 opcode uses; 0x59's second field is a literal 1 rather than
the total length symmetry with 0x56 Close predicts; and the checksum is a
zero-seeded CRC32 that is then inverted. The source document records the zero
seed and not the inversion, which is the likeliest single reason its multi-chunk
writes were refused on checksum.

Multi-chunk stays refused above 16 KiB rather than guessed. The captured file
fitted one chunk, so offset = 0 is all the evidence covers, and at one chunk a
byte offset and a sequence number are indistinguishable.

A file carries its own bank and slot at +23/+24 and the box rewrites them at
commit, so a correct copy to another slot differs in exactly two bytes.
differs_only_by_location reads that right and still catches a corruption or a
truncation. Measured three times.

Also here, both found by reading a whole project off the box for the first time
(2,061,057 bytes, ~503 chunks, and the first real exercise of the multi-chunk
read loop — every file that path was verified on before was a single-chunk
preset):

- file_declared_size read the payload length as a u16 at +27. It is a u32 at
  +25. All 32 committed fixtures agree under both readings because every one is
  a preset under 64 KB; the project is the first file that can tell them apart,
  where the old reading returns 29,398 instead of 2,061,014. The header now
  reads end to end with nothing spare.
- Writes are never retried. The generic request path re-sent each unanswered
  0x58 three times, which is right for an idempotent read and wrong for a write
  chunk.

+++ crates/app/examples/read_drive_preset.rs
@@ -55,7 +55,12 @@ fn main() {
                     // counted (checked inside `drive_read_file`). Agreement
                     // across all three is the claim being made here.
                     let declared = file_declared_size(&file);
-                    let agree = declared == Some(listed as u16);
+                    // Compared as `u32`, not `u16`. The cast that used to be
+                    // here truncated the *listing's* size to sixteen bits as
+                    // well, so both sides of this comparison lost the same high
+                    // half and agreed no matter what — the check could not have
+                    // failed on any file big enough to need it.
+                    let agree = declared == Some(listed);
                     println!(
                         "  {path:<18} {:>5}b read  header says {:?}  listing said {listed}  {}",
                         file.len(),
